chore: rename button component
