+++ ui/src/components/table/TableHeader.jsx
@@ -7,7 +7,7 @@ import { Typography } from '@splunk/react-ui/Typography';
 import styled from 'styled-components';
 import { _ } from '@splunk/ui-utils/i18n';
 
-import { UCCButton } from '../Button/Button';
+import { UCCButton } from '../UCCButton/UCCButton';
 import TableFilter from './TableFilter';
 import { TableSelectBoxWrapper } from './CustomTableStyle';
 import { PAGE_INPUT } from '../../constants/pages';
